# internal/fleet/server/openapi_types.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package server

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateFleet defines model for CreateFleet.
type CreateFleet struct {
	Capacity int    `json:"capacity" validate:"required,gt=0"`
	Name     string `json:"name" validate:"required"`
}

// CreateVehicleRequest defines model for CreateVehicleRequest.
type CreateVehicleRequest struct {
	Vin string `json:"vin" validate:"required,uuid"`
}

// DriverResponse defines model for DriverResponse.
type DriverResponse struct {
	Id string `json:"id"`
}

// Error defines model for Error.
type Error struct {
	Code    string  `json:"code"`
	Message *string `json:"message,omitempty"`
}

// FleetResponse defines model for FleetResponse.
type FleetResponse struct {
	Capacity int                     `json:"capacity"`
	Name     string                  `json:"name"`
	Vehicles *[]FleetVehicleResponse `json:"vehicles,omitempty"`
}

// FleetVehicleResponse defines model for FleetVehicleResponse.
type FleetVehicleResponse struct {
	Id *string `json:"id,omitempty"`
}

// IdResponse defines model for IdResponse.
type IdResponse struct {
	Id string `json:"id"`
}

// VehicleResponse defines model for VehicleResponse.
type VehicleResponse struct {
	Driver *DriverResponse `json:"driver,omitempty"`
	Id     string          `json:"id"`
	Vin    string          `json:"vin"`
}

// CreateFleetJSONRequestBody defines body for CreateFleet for application/json ContentType.
type CreateFleetJSONRequestBody = CreateFleet

// CreateVehicleJSONRequestBody defines body for CreateVehicle for application/json ContentType.
type CreateVehicleJSONRequestBody = CreateVehicleRequest